Add caching to file downloads
This should prevent multiple network requets for files. Instead if the
fileNode from `createRemoteFileNode` exists in the cache, we use
`touchNode` to tell Gatsby that the cached file should not be removed.

However if the item does not exist in the cache, then we download the
fileNode and update the KBFS cache to include the fileNodeId.